+++ packages/framer-motion/src/gestures/focus.ts
@@ -1,55 +1,49 @@
 import { addDomEvent } from "../events/use-dom-event"
+import { Feature } from "../motion/features/Feature"
 import { AnimationType } from "../render/utils/types"
-import type { VisualElement as MotionNode } from "../render/VisualElement"
 import { noop } from "../utils/noop"
 import { pipe } from "../utils/pipe"
 
-export function focus(node: MotionNode<Element>) {
-    let cleanEvents: Function = noop
-    let isFocusActive = false
+export class FocusGesture extends Feature<Element> {
+    private isActive = false
 
-    const onFocus = () => {
-        let isFocusVisible = false
+    private removeEventListeners: Function = noop
 
-        /**
-         * If this element doesn't match focus-visible then don't
-         * apply whileHover. But, if matches throws that focus-visible
-         * is not a valid selector then in that browser outline styles will be applied
-         * to the element by default and we want to match that behaviour with whileFocus.
-         */
-        try {
-            isFocusVisible = node.current!.matches(":focus-visible")
-        } catch (e) {
-            isFocusVisible = true
+    mount() {
+        const onFocus = () => {
+            let isFocusVisible = false
+
+            /**
+             * If this element doesn't match focus-visible then don't
+             * apply whileHover. But, if matches throws that focus-visible
+             * is not a valid selector then in that browser outline styles will be applied
+             * to the element by default and we want to match that behaviour with whileFocus.
+             */
+            try {
+                isFocusVisible = this.node.current!.matches(":focus-visible")
+            } catch (e) {
+                isFocusVisible = true
+            }
+
+            if (!isFocusVisible || !this.node.animationState) return
+
+            this.node.animationState.setActive(AnimationType.Focus, true)
+            this.isActive = true
         }
 
-        if (!isFocusVisible || !node.animationState) return
+        const onBlur = () => {
+            if (!this.isActive || !this.node.animationState) return
+            this.node.animationState.setActive(AnimationType.Focus, false)
+            this.isActive = false
+        }
 
-        node.animationState.setActive(AnimationType.Focus, true)
-        isFocusActive = true
+        this.removeEventListeners = pipe(
+            addDomEvent(this.node.current!, "focus", onFocus),
+            addDomEvent(this.node.current!, "blur", onBlur)
+        )
     }
 
-    const onBlur = () => {
-        if (!isFocusActive || !node.animationState) return
-        node.animationState.setActive(AnimationType.Focus, false)
-        isFocusActive = false
-    }
-
-    const cleanListeners = pipe(
-        node.once("Effect", () => {
-            cleanEvents = pipe(
-                addDomEvent(node.current!, "focus", onFocus),
-                addDomEvent(node.current!, "blur", onBlur)
-            )
-        }),
-        node.once("Unmount", () => {
-            cleanEvents()
-            cleanListeners()
-        })
-    )
-
-    return () => {
-        cleanEvents()
-        cleanListeners()
+    unmount() {
+        this.removeEventListeners()
     }
 }