+++ packages/framer-motion/src/motion/features/layout/index.ts
@@ -1,9 +1,0 @@
-import { FeatureComponents } from "../types"
-import { MeasureLayout } from "./MeasureLayout"
-
-export const layoutFeatures: FeatureComponents = {
-    measureLayout: {
-        type: "react",
-        feature: MeasureLayout,
-    },
-}